runtime: Move control flow related classes to separate file
This allows to use the constructs in other classes of the runtime.
Previously those were included in lib.h which would make this
impossible
